move hover-related classes to hover.py
